fix issues pointed out by no-sign-compare removal
By default comparing signed with unsigned, the compiler converts the
signed value to unsigned, this can cause problems sometimes, as that
would mean -1 > 2U.

In channel.c, context.c, device.c the lengths of strings to a negative
number, so cast to a signed number in all cases.

In dns_sd.c, local.c, tests/iio_stresstest.c we had an iterator, that
was signed, that should have been unsigned (now it is).

In local.c:get_rel_timeout_ms we move the comparision to signed.

in local.c, while pointers are unsigned integers, subtraction of two
pointers are ptrdiff_t, or a signed integer. Since we are just measuring
length of a string, cast to unsigned.

+++ channel.c
@@ -293,14 +293,14 @@ char * iio_channel_get_xml(const struct iio_channel *chn, size_t *length)
 		len = eptr - ptr;
 	}
 
-	if (chn->is_scan_element && len > scan_element_len) {
+	if (chn->is_scan_element && len > (ssize_t) scan_element_len) {
 		memcpy(ptr, scan_element, scan_element_len); /* Flawfinder: ignore */
 		ptr += scan_element_len;
 		len -= scan_element_len;
 	}
 
 	for (i = 0; i < chn->nb_attrs; i++) {
-		if (len > attrs_len[i]) {
+		if (len > (ssize_t) attrs_len[i]) {
 			memcpy(ptr, attrs[i], attrs_len[i]); /* Flawfinder: ignore */
 			ptr += attrs_len[i];
 			len -= attrs_len[i];
